Starting point for testing multi-server messages.
This takes the approach used for homework #2 and strips out
implementations of specific commands (except for a new
"sample" command that shows how arbitrary commands might be
constructed).  The sample command simply proposes a value
that is then replicated to the CommonState variables of all
other servers.

The script "run_tests.sh" can be used to launch servers and
clients and send them a series of test commands.  Servers
are configured with the "input-srv<N>.txt" files in the way
suggested for homework #2.

The Java code can be built using the Makefile but it should
also be possible to build it in other environments.